+++ client/app/common/services/localStorageService.js
@@ -1,0 +1,34 @@
+/* Copyright (c) Trainline Limited, 2016-2017. All rights reserved. See LICENSE.txt in the project root for license information. */
+
+'use strict';
+
+(function () {
+  angular
+    .module('EnvironmentManager.common')
+    .factory('localstorageservice', localstorageservice);
+
+  localstorageservice.$inject = [];
+
+  function localstorageservice() {
+    var localStorage = window.localStorage;
+
+    return {
+      get: get,
+      set: set,
+      exists: exists
+    };
+
+    function get(key) {
+      return localStorage.getItem(key);
+    }
+
+    function set(key, value) {
+      return localStorage.setItem(key, value);
+    }
+
+    function exists(key) {
+      if (get(key) === null) return false;
+      return true;
+    }
+  }
+}());